requirements.txt
altmetric==1.0
biopython==1.70
certifi==2018.1.18
chardet==3.0.4
colorlog==3.1.2
idna==2.6
numpy==1.14.2
progressbar2==3.36.0
PubMedMetrics==1.0.0
python-utils==2.3.0
PyYAML==3.12
requests==2.18.4
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.5
unicodecsv==0.14.1
urllib3==1.22
